requirement.txt
ultralytics
torch
torchvision
onnx
onnx2tf
tensorflow-cpu  # 导出用